Updated the schema types from int to float

+++ graphql/typeDefs.ts
@@ -66,13 +66,13 @@ export const typeDefs = gql` #graghql
         popularity:Int
         id:ID!
         backdrop_path:String
-        vote_average:Int
+        vote_average:Float
         overview:String
         first_air_date:String
         origin_country:[String]
         genre_ids:[Int]
         original_language:String
-        vote_count:Int
+        vote_count:Float
         name:String
         original_string:String
     }
